+++ packages/eslint-config-drivers/jest/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  collectCoverageFrom: ['<rootDir>/src/**/*.{ts,js}'],
+};