+++ packages/eslint-config-drivers/test/a.ts
@@ -1,0 +1,5 @@
+/* eslint-disable no-console */
+
+const func = (): void => console.log('A');
+
+export default func;